progress,tools: move CopyCallback (and related) to 'tools'
